Function rustc_symbol_mangling::symbol_name_for_instance_in_crate

pub fn symbol_name_for_instance_in_crate<'tcx>(
    tcx: TyCtxt<'tcx>,
    instance: Instance<'tcx>,
    instantiating_crate: CrateNum,
) -> String

This function computes the symbol name for the given instance and the given instantiating crate. That is, if you know that instance X is instantiated in crate Y, this is the symbol name this instance would have.